+++ extensions/spotify-player/src/helpers/formatTitle.ts
@@ -1,7 +1,19 @@
-export function formatTitle(title: string, max: number, showEllipsis = true) {
+export function formatTitle(
+  name: string | undefined,
+  artist: string | undefined,
+  max: number,
+  showArtist: boolean,
+  showEllipsis = true,
+) {
   if (max === 0) {
     return "";
   }
+
+  if (!name || !artist) {
+    return "";
+  }
+
+  const title = showArtist ? `${name} - ${artist}` : name;
 
   if (title.length <= max) {
     return title;